code refractor of Medications page
